Prevent duplicate downloads if script is re-ran

+++ cloud/deploy-local.sh
@@ -2,8 +2,10 @@
 
 set -e
 
-# Pull in the docker-compose template & variables
-wget -O docker-compose.yaml "https://github.com/MORpheus-Software/NFA/blob/main/UniversalBuilder/lib/docker-compose.yaml"
+# Pull in the docker-compose template if not present
+if [ ! -f docker-compose.yaml ]; then
+    wget -O docker-compose.yaml "https://github.com/MORpheus-Software/NFA/blob/main/UniversalBuilder/lib/docker-compose.yaml"
+fi
 
 # Check if the env file exists, otherwise rename the example file
 if [ ! -f .env ]; then
